+++ modules/apim-adapter/src/main/java/com/axway/apim/adapter/apis/APIManagerOrganizationAdapter.java
@@ -1,5 +1,6 @@
 package com.axway.apim.adapter.apis;
 
+import java.io.InputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.HashMap;
@@ -137,13 +138,13 @@ public class APIManagerOrganizationAdapter {
 				RestAPICall request;
 				if(actualOrg==null) {
 					String json = mapper.writeValueAsString(desiredOrg);
-					HttpEntity entity = new StringEntity(json);
+					HttpEntity entity = new StringEntity(json, ContentType.APPLICATION_JSON);
 					request = new POSTRequest(entity, uri, true);
 				} else {
 					desiredOrg.setId(actualOrg.getId());
 					if (desiredOrg.getDn()==null) desiredOrg.setDn(actualOrg.getDn());
 					String json = mapper.writeValueAsString(desiredOrg);
-					HttpEntity entity = new StringEntity(json);
+					HttpEntity entity = new StringEntity(json, ContentType.APPLICATION_JSON);
 					request = new PUTRequest(entity, uri, true);
 				}
 				request.setContentType("application/json");
@@ -196,8 +197,9 @@ public class APIManagerOrganizationAdapter {
 		if(actualOrg!=null && org.getImage().equals(actualOrg.getImage())) return;
 		HttpResponse httpResponse = null;
 		URI uri = new URIBuilder(cmd.getAPIManagerURL()).setPath(RestAPICall.API_VERSION+"/organizations/"+org.getId()+"/image").build();
+		InputStream is = org.getImage().getInputStream();
 		HttpEntity entity = MultipartEntityBuilder.create()
-			.addBinaryBody("file", org.getImage().getInputStream(), ContentType.create("image/jpeg"), org.getImage().getBaseFilename())
+			.addBinaryBody("file", is, ContentType.create("image/jpeg"), org.getImage().getBaseFilename())
 			.build();
 		try {
 			RestAPICall apiCall = new POSTRequest(entity, uri);
@@ -213,6 +215,9 @@ public class APIManagerOrganizationAdapter {
 			try {
 				((CloseableHttpResponse)httpResponse).close();
 			} catch (Exception ignore) { }
+			try {
+				is.close();
+			} catch (Exception ignore) { } 
 		}
 	}